# Conflicts:
#	mmdet/models/detectors/two_stage.py

+++ mmdet/core/visualization/image.py
@@ -76,7 +76,7 @@ def imshow_det_bboxes(img,
         'bboxes.shape[0] and labels.shape[0] should have the same length.'
     assert bboxes.shape[1] == 4 or bboxes.shape[1] == 5, \
         f' bboxes.shape[1] should be 4 or 5, but its {bboxes.shape[1]}.'
-    img = mmcv.imread(img)
+    img = mmcv.imread(img).astype(np.uint8)
 
     if score_thr > 0:
         assert bboxes.shape[1] == 5
